requirements.txt
annotated-types==0.7.0
appier==1.34.5
beautifulsoup4==4.12.3
boto3==1.35.70
botocore==1.35.70
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
docstring_parser==0.16
faiss-cpu==1.9.0.post1
filelock==3.16.1
fsspec==2024.10.0
google==3.0.0
google-api==0.1.12
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.73.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.13.1
grpcio==1.68.0
grpcio-status==1.68.0
httplib2==0.22.0
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.2
pydantic_core==2.27.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
s3transfer==0.10.4
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.6.0
shapely==2.0.6
six==1.16.0
soupsieve==2.6
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.6.0.dev20241126
torchaudio==2.5.0.dev20241126
torchvision==0.20.0.dev20241126
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
